Time: 243 ms (5.07%), Space: 108.6 MB (27.43%) - LeetHub

+++ 3075-maximize-happiness-of-selected-children/3075-maximize-happiness-of-selected-children.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     long long maximumHappinessSum(vector<int>& happiness, int k) {
         sort(happiness.rbegin(),happiness.rend());
-        long long n=happiness.size(),res=0;
-        for(long long i=0;i<n;i++){
+         long long n=happiness.size(),res=0;
+        for(int i=0;i<n;i++){
             if(k==0) return res;
             if((happiness[i]-i)<0) res=res+0;
             else res=res+(happiness[i]-i);
